+++ src/test/java/org/scion/jpan/api/DatagramSocketApiTest.java
@@ -347,7 +347,7 @@ class DatagramSocketApiTest {
   @Test
   void getService_default() throws IOException {
     ScionService service1 = Scion.defaultService();
-    ScionService service2 = Scion.newServiceWithDaemon(MockDaemon.DEFAULT_ADDRESS_STR);
+    ScionService service2 = Scion.newServiceWithDaemon(MockDaemon.getAddressStr());
     try (ScionDatagramSocket socket = new ScionDatagramSocket()) {
       assertEquals(service1, socket.getService());
 
@@ -362,7 +362,7 @@ class DatagramSocketApiTest {
   @Test
   void getService_non_default() throws IOException {
     ScionService service1 = Scion.defaultService();
-    ScionService service2 = Scion.newServiceWithDaemon(MockDaemon.DEFAULT_ADDRESS_STR);
+    ScionService service2 = Scion.newServiceWithDaemon(MockDaemon.getAddressStr());
     try (ScionDatagramSocket socket = ScionDatagramSocket.newBuilder().service(service2).open()) {
       assertEquals(service2, socket.getService());
       assertNotEquals(service1, socket.getService());