+++ CHANGELOG.md
@@ -191,6 +191,10 @@ TODO
   [#281](https://github.com/scionproto-contrib/jpan/pull/281)
 - Closeable PathSelectors and close() calling SimpleCache
   [#280](https://github.com/scionproto-contrib/jpan/pull/280)
+- Added IPv6 only topology + tests
+  [#282](https://github.com/scionproto-contrib/jpan/pull/282)
+TODO
+- remove DemoTopology?
 
 ### Fixed